update farmer service/serviceimpl

+++ src/main/java/com/farmily/user/model/FarmerReview.java
@@ -29,7 +29,7 @@ public class FarmerReview implements Serializable {
     private Admin admin;
 
     @Column(name = "review_round", nullable = false)
-    private Byte reviewRound = 1;        // 預設初始值 = 1
+    private Integer reviewRound = 1;        // 預設初始值 = 1
 
     @Enumerated(EnumType.STRING)
     @Column(name = "review_status")
@@ -99,11 +99,11 @@ public class FarmerReview implements Serializable {
         this.admin = admin;
     }
 
-    public Byte getReviewRound() {
+    public Integer getReviewRound() {
         return reviewRound;
     }
 
-    public void setReviewRound(Byte reviewRound) {
+    public void setReviewRound(Integer reviewRound) {
         this.reviewRound = reviewRound;
     }
